refactor(MoveResult): controller 패키지로 이동
